proposal for GUI improvements
Within the main updates, both within the Localize and Render tools there are:
- the cursor x, y positions as well as their values are displayed within the status bar
- the mouse wheel zooms/unzooms centering on the mouse cursor position
- it is as well possible to zoom by simply drawing a rectangle
- as a left double-clicking is fully unzooming, i.e. similar to a View>Fit_image_to_window or Ctrl+w
- and a right click and move able to drag the window in the case it is zoomed (with the open hand mouse cursor)
Within the Render tool it is as well possible to set a pick with a Ctrl+Mouse_left_double_click even when being within the zoom mode.
As within the Localize tool it is possible to draw a ROI by pressing the cursor key and drawing it.
By making a mouse right click inside the ROI and moving it, the ROI will be displaced (with the closed hand mouse cursor) and outside the ROI it is the window that is displaced (with the open hand mouse cursor as described higher).

Within the render tool, I added the possibility to enter/leave the 3D slicing mode with the Ctrl+Q shortkeys.
Once in the 3D slicing mode, the given slice position (and size) is now indicated within the status_bar_frame_indicator and it is possible to change the slice position with a Ctrl+mouse_wheel.
I also changed to "3D slicer menu" the original menu link within which I increased the size of the plot so that it is fully fitting within the window.
What I had as well implemented previously and forgot to describe you is that it is possible to add a new pick with a Ctrl+left_mouse_double_click (within the Zoom mode) and erase a given pick with a Ctrl+right_mouse_double_click